refactor: extract JWSEncoder class
